+++ src/main/java/net/fabricmc/loader/api/SemanticVersion.java
@@ -17,7 +17,6 @@
 package net.fabricmc.loader.api;
 
 import net.fabricmc.loader.util.version.VersionDeserializer;
-import net.fabricmc.loader.util.version.VersionParsingException;
 
 import java.util.Optional;
