+++ src/main/java/net/fabricmc/loader/metadata/BuiltinModMetadata.java
@@ -26,13 +26,13 @@ import java.util.Optional;
 import java.util.TreeMap;
 
 import net.fabricmc.loader.api.Version;
+import net.fabricmc.loader.api.VersionParsingException;
 import net.fabricmc.loader.api.metadata.ContactInformation;
 import net.fabricmc.loader.api.metadata.CustomValue;
 import net.fabricmc.loader.api.metadata.ModDependency;
 import net.fabricmc.loader.api.metadata.ModMetadata;
 import net.fabricmc.loader.api.metadata.Person;
 import net.fabricmc.loader.util.version.VersionDeserializer;
-import net.fabricmc.loader.util.version.VersionParsingException;
 
 public final class BuiltinModMetadata implements ModMetadata {
 	private final String id;
